requirements.txt
numpy
pandas
h5py
Pillow
opencv-python
scipy
matplotlib
tqdm